# docs/swagger.yaml
basePath: /v1/api
definitions:
...
  CONVERDA_internal_notification_controller_dto.CreateGroupRequest:
    properties:
      description:
        type: string
      is_default:
        type: boolean
      key:
        type: string
      name:
        type: string
    required:
    - key
    - name
    type: object
  CONVERDA_internal_notification_controller_dto.CreateLayoutRequest:
    properties:
      content_html:
        type: string
      description:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      variables_schema:
        additionalProperties: true
        type: object
    required:
    - content_html
    - name
    type: object
  CONVERDA_internal_notification_controller_dto.GroupResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      key:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  CONVERDA_internal_notification_controller_dto.LayoutResponse:
    properties:
      content_html:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      variables_schema:
        additionalProperties: true
        type: object
    type: object
  CONVERDA_internal_notification_controller_dto.NotificationResponse:
    properties:
      notification_id:
        type: string
      status:
        type: string
    type: object
  CONVERDA_internal_notification_controller_dto.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  CONVERDA_internal_notification_controller_dto.SendNotificationRequest:
    properties:
      channel:
        enum:
        - email
        - sms
        - push
        type: string
      data:
        additionalProperties: true
        type: object
      language:
        description: Optional, default 'en'
        type: string
      recipient:
        description: Validate email format
        type: string
      template_code:
        type: string
    required:
    - channel
    - recipient
    - template_code
    type: object
...
  CONVERDA_internal_notification_controller_dto.UpdateLayoutRequest:
    properties:
      content_html:
        type: string
      description:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      variables_schema:
        additionalProperties: true
        type: object
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a server for a Go Drunk Backend API, demonstrating DDD principles.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Drunk Backend API by DDD
  version: "1.0"
paths:
...
  /notifications/groups:
    get:
      description: List notification groups
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: List Notification Groups
      tags:
      - Notification
    post:
      consumes:
      - application/json
      description: Create a new notification group (category)
      parameters:
      - description: Group Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.GroupResponse'
      security:
      - BearerAuth: []
      summary: Create Notification Group
      tags:
      - Notification
...
  /notifications/layouts:
    get:
      description: List notification layouts
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: List Notification Layouts
      tags:
      - Notification
    post:
      consumes:
      - application/json
      description: Create a new notification layout
      parameters:
      - description: Layout Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.CreateLayoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.LayoutResponse'
      security:
      - BearerAuth: []
      summary: Create Notification Layout
      tags:
      - Notification
  /notifications/layouts/{id}:
    delete:
      description: Delete a notification layout
      parameters:
      - description: Layout ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Delete Notification Layout
      tags:
      - Notification
    get:
      description: Get a notification layout by ID
      parameters:
      - description: Layout ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.LayoutResponse'
      security:
      - BearerAuth: []
      summary: Get Notification Layout
      tags:
      - Notification
    patch:
      consumes:
      - application/json
      description: Update a notification layout
      parameters:
      - description: Layout ID
        in: path
        name: id
        required: true
        type: string
      - description: Layout Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.UpdateLayoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Update Notification Layout
      tags:
      - Notification
  /notifications/send:
    post:
      consumes:
      - application/json
      description: Send a notification via configured channel
      parameters:
      - description: Notification Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.SendNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.NotificationResponse'
      security:
      - BearerAuth: []
      summary: Send Notification
      tags:
      - Notification
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"